refactor LoginDAO
